move functions into ssh_parse.go
