+++ age/tui.go
@@ -1,5 +1,6 @@
 // These functions have been copied from the age project
 // https://github.com/FiloSottile/age/blob/v1.0.0/cmd/age/encrypted_keys.go
+//
 // Copyright 2021 The age Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in age's LICENSE file at
